Loosen restriction for building deprecated iOS targets
This patch loosens the restriction for building deprecated iOS targets
to only when the user attempts to compile a binary or emit IR.
This allows the compiler to continue building swift modules and
interfaces for newer deployment targets.

The error message is also slightly modified to allow the same message to
apply to different platforms.

resolves: <rdar://87898177>