SwiftDriver: change the response file emission
This adjusts the response file emission to follow the clang behaviour.
Doing so allows the use of response files on Windows to be passed to
clang, which will default to POSIX style response files unless
explicitly passed `--rsp-quoting=windows`.  This allows for a single
path for the emission across platforms, and avoids having to deal with
the dynamic switching of the response file formats.  With this change,
ignoring the serialization issue, it is now possible to build
swift-driver with swift-driver.