SwiftDriver: migrate and alter TSCUtility.dl*
This migrates the TSCUtility helpers for the loader to swift-driver in
order to enable deprecating the utility functions in tools-support-core.
This family of functions interacts with the dynamic linker (aka loader)
to loader and unload modules and lookup symbols by name.  The migration
takes the opportunity to adjust the API surface a small amount to read
more naturally in Swift rather than being a wrapper layer over the Unix
C APIs.