SwiftDriver: use `clang++` as linker driver with C++ Interop
This adjusts the linker job construction to switch the linker driver to
the C++ compiler driver as the linker driver when C++ interop is
enabled.  This allows us to ensure that we properly link against the C++
runtime which should be assumed to be needed with C++ interop.  This is
required for further exploration of C++ interop.