Track serialized diagnostics of emit-module jobs separately.
When serialized diagnostics are requested by the build system, it is expected to provide an output-file-map to specify per-file diagnostic output paths. The build system may also request serialized diagnostics module-wide in WMO compilation by specifying an output-file-map entry of module scope (`""`).

Emit-module jobs are always run at module scope, which means that if we want to get serialized diagnostics from the emit-module job, in WMO we do not know which path to use, because the module-scope path is specified for the WMO compilation job.

This change introduces a separate output path for the emit-module serialized diagnostics to be specified either:
- On the command-line with `-emit-module-serialize-diagnostics-path`
- In the output file-map using key `"emit-module-diagnostics": "…/…emit-module.dia"` at global (module) scope.

Resolves rdar://83110325