Driver: further generalise #1088
It seems that a similar problem has impacted the generation of
serialised diagnostics for modules.  Follow the adage of CS inherently
having 3 numbers: 0, 1, N and generalise the path to a penultimate
fallback.  We should always have the swift dependency information on
hand to locate the build directory as that is the "master dependency".
This should hopefully prevent accidental cases of files leaking into the
current directory.