Remove Noisy Remark
This remark is basically tripped by every swiftmodule in the SDK for
most projects. As it doesn't appear in tests anywhere and it provides
very little signal, best to strike it for now. If we discover it has
some diagnostic utility we can always revert this and add some tests.