SwiftDriverTests: adjust module count on Windows
A recent refactoring for Windows increased the module count for the
system modules and this resulted in the test suite failing due to the
difference.  This is technically backwards incompatible and expects that
the most recent toolchain is in use but it seems unclear how to account
for that easily.