Add ability to override/disable response file use in ArgsResolver
This change deprecates 'ArgsResolver' API that takes a boolean 'forceResponseFiles' in favour of a new API that takes an 'ResponseFileHandling' enum. This enum allows the previous behaviors of '.forced' and '.heuristic' ('true' and 'false', respectively of 'forceResponseFiles'), and adds an option to override disable use of response files completely ('.disabled').

This is useful in the example of dependency scanning when it is done via a C interface to a compiler library: even for exceedingly long command-line argument lists, using response files is not necessary since we just pass in a pointer to an already-allocated memory buffer containing the argument list, directly.

Resolves rdar://99296444