SwiftDriverTests: exclude explicit triple
The explicit triple sets an OS, which then requires that the value for
`-sdk` match.  Because Windows uses a `-sdk` parameter, this will fail
on Windows due to a mismatch for the SDKs.  By excluding the triple, a
host specific value is used which allows this to match and pass.