[Dependency Scanning] Adapt to upcoming dependency scanner API changes with the output path.
https://github.com/apple/swift/pull/60882 adds functionality to the Clang scan-deps invocation that produces placeholder paths as '<replace-me>'s in the generated command-line.
The first step to adapting to these changes is to teach the driver to substitute the expected file paths.

Once the scanner changes land, we will revisit the driver specifying these paths in favour of having the scanning action being able to determine the correct output location from the get-go.