SwiftDriver: migrate LLVM bitcode support from tools-support-core
The swift-driver is the only consumer of the bitcode implementation in
TSCUtility.  Migrate the code to the single user to allow us to
deprecate the API in tools-support-core.  While there are no practical
changes to the implementation, this does take the opportunity to split
up the bitcode implementation into multiple files.