Enable swiftinterface verification with compatiblity headers
Lift the workaround that disabled scheduling the swiftinterface
verification task when a compatibility header is generated. This is an
issue only when the header is also installed, we now expect the build
system to disable the verification when it installs the header.

This issue lead to errors on missing headers and failures to build the
underlying clang module at verification.

rdar://100989723