Use the versioned triple outside of Darwin.
Some platforms require the normal, versioned triple to be supplied for
the target flag, as Swift modules use this versioned triple as a
subdirectory. If the unversioned triple is used instead, the standard
library won't be found on these platforms.

Instead, use the unversioned triple only on Darwin. This also should be
a noop on Linux. This behavior matches apple/swift-package-manager#3576
and apple/swift-tools-support-core#229 which contain additional context.