[Dependency Scanning] Do not specify placeholder modules by default as inputs to scanning actions
Instead allow the dependency scanning oracle to be configured such that the scanning actions use placeholders.

In contexts where the driver's client can guarantee that all module dependencies have been built prior to planning of the dependee, placeholders are not necessary in the scanner, becuase the binary modules will simply be discovered on the filesystem.

Resolves rdar://102339586