[Dependency Scanning] Move computation of the path for libSwiftScan to the toolchain
Alongside lookup of executable tools, the toolchain is now responsible for looking up the driver's compiler support and dependency scanning library.

It also adds an environment variable override 'SWIFT_DRIVER_SWIFTSCAN_LIB' for clients to specify a specific path to this shared library that they would like to be used instead of the default.