Be more strict about invalid -module-name arguments
In the case this argument is passed explicitly and invalid, it was
previous ignored and fell back to 'main' even in cases where you might
not want that like compiling object files. It doesn't seem necessary to
allow invalid cases like this through regardless of what object type is
being built.