Add -gcc-toolchain flag to swift-driver
GCC toolchains are useful when cross-compiling to Linux with a given gcc
toolchain. The gcc toolchain contains tools for linking, libraries for
linking against, and the headers that map to those libraries.

This patch adds a gcc-toolchain flag to the driver so that Swift can
pass that down to both the clang importer and clang linker
appropriately.