[ELF] Don't pass ELF shared objects to autolink-extract
This is a version of apple/swift#30465 for the new driver.

Since shared objects are classified as `.object`, they end up getting
passed to the autolink-extract job.

For ELF platforms, this is unnecessary, since the autolink entries are
not in the final shared object, so the autolink-extract process will not
find any entries.

Additionally, in ELF platforms, `.so` files can actually be linker
scripts (text files) with instructions for the actual linker, which the
autolink-extract job will not understand and will fail.

For those two reasons, modify the autolink-extract inputs to skip over
objects in ELF platforms that have the `.so` extension.