Distro-overridable default Unix linker
Being able to specify various defaults that the driver should assume is
necessary for building specialized distributions of the toolchain. This
mostly affects the non-Darwin Unix environment, where folks may be using
a whole smorgasbord of different linkers in their toolchain. The
original experience made it impossible to generally build the driver to
default to a specific linker without editing source, which isn't great.

Swift has the challenge of not allowing values assigned to macro
definitions though, and neither does SwiftPM. Unless you have some fun
with the clang importer, of course :). So we introduce a new
DriverDefaults module that really just contains strings that were set as
defaults. It seems that that clang importer was unable to import macros
directly if they involved more than the literal, so I was unable to use
`SWIFT_DEFAULT_LINKER` directly since I want to quote it, but instead
have to assign it to an actual value.

To configure the default linker while building with SwiftPM,
`swift build -Xcc -DSWIFT_DEFAULT_LINKER=lld`, and you will end up with
a Swift driver that will default to using lld on Linux.