build: enable building in Debug mode on Windows
Windows does not have an ABI stable C/C++ runtime, and is ABI
incompatible between debug and release mode on the same version even.
Ensure that we always use the release mode DLL variant of the C/C++
runtime (`/MD`) allowing us to build the client code in debug or release
mode.