build: support building swift-driver outside of the toolchain
We can now vendor the dependencies allowing us to build a separate copy
of the dependencies when building the package if it is not provided for.
This enables building the package standalone for debugging in scenarios
where SPM is unable to build the package (e.g. due to incorrect
linking).