[CAS] Don't create the same CAS multiple times from the same Oracle
If the oracle has been used to create a CAS for the path, reuse the CAS
if it is asked to create the same CAS again. This not only save the work
of creating the CAS, but also avoid a potential problem that the CAS
closing synchronization doesn't work across threads in the same process
due to its uses of file lock.