Fix explicit modules incremental build failure when only the clang module of a mixed-source module requires rebuild
After reading a prior intermodule dependency graph in an incremental explicit module build, the driver loops over
each module in the graph to check if it is up to date. The check which is intended to skip the main (Swift) module
was including a clang module with the same name as well, if it existed. If this clang module was the only one to be
updated in an incremental compile, we wouldn't regenerate its pcm, and compilation would later fail since it wasn't
up to date.

rdar://121352768