Execution: special case the first two arguments in RSP mode
The second argument is often treated specially as a mode specifier for
tools. This is particularly important for Windows where `-lib` or `/lib`
passed to `link` or `ld.lld-link` will change the linker to the
librarian. This argument may not be sunk into the response file as it is
not treated as the mode specifier then. Special case the first two
arguments to ensure that the mode switch parameter is always passed for
any tools.