Update Generic Unix linker selection
Not all generic-"unix" environments have the Gold linker available to
them, and in some cases, the vendor of the toolchain may provide their
own linker. In these cases, the driver should be internally consistent
with the toolchain that it is shipped with.

Now that we have the clang-linker, we can lean on the linker selection
in the clang-linker to determine a default linker. If the clang-linker,
and thus, the swift compiler driver, are part of a specific toolchain,
that clang-linker should be built for that platform with the appropriate
linker defaults set. If someone overrides the linker with
`-use-ld`, we should still honour that, but should otherwise be
consistent with the appropriate toolchain linker.

Fixes: rdar://123061492