[unix] Fix linker argument in Unix toolchains
The spelling of the linker is `-fuse-ld=`, not `--fuse-ld=`, as it can
be seen in [1] and as used in the C++ driver in [2].

The code was correct before #1545 using `-fuse-ld=`, but was changed to
`--fuse-ld=`. The test was testing for the command line containing the
same spelling, so it would had never failed, but running the actual
command line would had failed. That code was eventually reverted and
reverted back in #1608.

[1]: https://github.com/apple/llvm-project/blob/e33819dff1cbf7a90b0216a4993126cb11440d45/clang/include/clang/Driver/Options.td#L5267
[2]: https://github.com/apple/swift/blob/319f36b01baa082fd2d2640d6e320594752702d0/lib/Driver/UnixToolChains.cpp#L234