[Explicit Module Builds] Emit scanner-global diagnostics on a total dependency scan failure
The driver queries per-scan diagnostics when the scanner query produces a result that is incomplete or invalid. When the scanner query completely fails to produce any kind of output, it may still be useful to fallback onto the global scanner diagnostic state.

The more forward-looking solution is to ensure scanner queries *always* produce enough state to be able to diagnose their failure.

Resolves rdar://130897498