Pass `-lld-allow-duplicate-weak` for coverage on Windows
Profiling currently relies on the ability to emit duplicate weak
symbols across translation units and having the linker coalesce them.
Unfortunately `link.exe` does not support this, so require `lld-link`
and pass `-lld-allow-duplicate-weak` when `-profile-generate` is
passed. We ought to be able to remove this once coverage is changed
to not rely on duplicate weak symbols, rdar://131295678 is tracking
that.

rdar://129337999