[Diagnostics] Support for all the flags of the warning treating group
This commit adds support for the warning treating option group, including the following options: -warnings-as-errors, -no-warnings-as-errors, -Werror, and -Wwarning.
Options in this group are now preserved as-is. It is forbidden to reorder or drop any of them.
These changes reflect the modifications made to the frontend in https://github.com/swiftlang/swift/pull/74466.