SwiftDriver: forward `-sysroot` provided by the user
This addresses a missing component of the initial Android support. When
the user specifies `-sysroot`, that should always be given precedence as
it is explicitly specified by the user. The `ANDROID_NDK_ROOT`
environment variable is set by the Android NDK and is used as a
_default_ value in the scenario that the user does not specify the
`-sysroot` option. This makes Android both behave more like the Windows
platform and also ensures that the user has full control over the
behaviour of the toolchain.

Take the opportunity to refactor some of the code to extract the Android
NDK specific helpers into an uninhabited enum and provide some
behavioural test cases.