Tests: adjust tests for Darwin
The Darwin toolchain is incomplete, missing `swift-autolink-extract`.
Add a workaround for the missing tool to allow the tests to pass. This
then uncovers another issue with ASi builds where the prebuilt NDK is
only available on x86_64 hosts. While we could loosen the NDK lookup, we
do still occassionally use some of the tools, so for now lets limit the
test to x86_64.

Fixes: #1682