[Explicit Module Builds] Always filter out up-to-date module dependencies
With an escape hatch of '-always-rebuild-module-dependencies'.
This change makes it so the driver is always going to produce a build plan which contains only explicit module pre-compile tasks for modules which are not already up-to-date with respect to all of their inputs.

This means that the driver will skip up-to-date module tasks even when '-incremental' is not specified. This is more in line with prior behavior when using Implicit Module builds and results in a better user experience when comppiling code with `swiftc` on the command-line.

For example, consequtive invocations of 'swiftc test.swift -o test.out' should not have to re-build possible module-dependencies of 'test.swift' each time.