swift-driver: correct the legacy driver path construction
Use `URL` to compute the new executable name. This is important to
ensure that we correctly strip the executable suffix and re-apply it in
the proper location if applicable. Without this change, on Windows, the
fallback path would generate `swiftc.exe-legacy-driver` rather than
`swiftc-legacy-driver.exe` for the basename of the C++ driver.