SwiftDriver: adjust handling of `-resource-dir` parameter
When computing the resource directory for the Swift registrar, we need
to add a couple of components that we had missed. Correct the path
computation so that we can find the object file properly. This was
detected while running the test suite with the early swift driver on
Windows where `-resource-dir` is explicitly passed.