Disable in-process target-info and supported-features queries using libSwiftScan
When building the compiler toolchain, there is currently a race condition which makes it so that the 'libSwiftScan.dylib' is not ready by the time the driver is used to begin compilation of the Swift standard library. For now, disable in-process queries altogether.

On 'main', this is resolved with https://github.com/swiftlang/swift/pull/77606, which seems to be a challenge to backport to 6.0.