Tests: introduce and partially adopt a command line test helper
Introduce a `XCTAssertCommandLineContains` and
`XCTAssertJobInvocationMatches` to validate the invocation of the tools.
This helper checks that the command line matches the template and in the
case of a failure will print both the actual command line template and
the expectation so that it is easier to diagnose the failures.

Take the opportunity to clean up some of the job type checks, converting
them from `XCTAssertTrue` to `XCTAssertEquals` to ensure that the values
are printed.

Perform a partial pass to adopt this new helper in a large set of the
SwiftDriverTests. This is a mostly mechanical change, though there are a
few places where the tests are actually changed semantically slightly.
Those changes are limited to a stronger assertion about the ordering of
arguments which was not previously present. These are options where file
paths are specified as an argument to an option.