Fix the legacy driver path construction on non-Windows platforms
`swift-driver` tries to invoke `swiftc-legacy-driver.` (trailing dot)
on non-Windows platforms when legacy driver is requested since
https://github.com/swiftlang/swift-driver/pull/1746

This is because `URL.pathExtension` returns an empty string if the
URL does not have an extension but `URL.appendPathExtension` appends
a period before the extension even if given an empty string.
Thus `url.deletingPathExtension().appendingPathExtension(url.pathExtension)`
will not be equal to `url` if `url` does not have an extension.