Group supplementary module outputs
This should have no actual impact on compilation. The idea is that we
are grouping the supplementary outputs that are specific to a given
target. This will allow us to store two of them, one for the target and
one for the target-variant. The end goal being that we can emit the
target variant interface bits in the same driver invocation that
generates the zippered object files.