tests: silence debug output, simplify and tighten checks
Use the `assertDriverDiagnostics` helper to ensure that all the emitted
diagnostics are verified. Remove the debug printing and output to the
logs. This is no longer necessary as we will verify that any additional
diagnostics emitted are checked and the helper will render the missed or
non-matching diagnostics allowing easy triaging of the failures in CI.